Packy and i3theme updates
Packy gets a small update, it handles better makepkg errors. Now it detects if makepkg exits with no errors and if the package is already built, useful if sudo password expires.
i3theme now better detects the theme dependencies. It checks if the package is installed, and if not, check if its a theme and if its installed on /usr/share/themes.